Sortup code. Switch range when inverted
